Highscore done. Working on level editor.
Already put the editor option in the main menu. Created the interface like level select. Will have to make copied of all levels on it and link the buttons to those levels to make it accessible to the player. Need to create a whole scene altogether to create new levels from scratch. Buttons that we'll need:
starting point tile, tiles, strong switch, normal switch, teleswitch, bridge.

Need to make button to toggle the timer as the player needs to select or unselect it.

Need to make bonus level.

Need to put music.